Add gestore errore shop

+++ controllers/shop/item.js
@@ -11,8 +11,6 @@ const fs = require("fs");
 const utils = require("../../utilities");
 const error = require("../../error_handler");
 
-const MISSING_CATEGORY_MESSAGE = "Categoria non esistente";
-
 /* 
     Gestisce la creazione di un nuovo item comprensivo di prodotti 
 */
@@ -26,7 +24,6 @@ async function createItem(req, res) {
 
         // Estrazione id della categoria
         const category = await CategoryModel.findByName(to_insert_item.category);
-        if (!category) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(MISSING_CATEGORY_MESSAGE)); }
         delete to_insert_item.category;
         to_insert_item.category_id = category._id;
 
@@ -41,13 +38,11 @@ async function createItem(req, res) {
         new_item_id = new_item.id;
     }
     catch (err) {
-        switch (err.code) {
-            case utils.MONGO_DUPLICATED_KEY:
-                await ProductModel.deleteMany({ _id: products_id }).catch((err) => {}); // Rimuove tutti i prodotti dato che non verranno associati all'item
-                return res.status(utils.http.CONFLICT).json({ field: "barcode", message: "Il prodotto associato al barcode è già presente in un item" });
-            default:
-                return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        if (err.code == utils.MONGO_DUPLICATED_KEY) {
+            await ProductModel.deleteMany({ _id: products_id }).catch((err) => {}); // Rimuove tutti i prodotti dato che non verranno associati all'item
+            err = error.generate.CONFLICT({ field: "barcode", message: "Il prodotto associato al barcode è già presente in un item" });
         }
+        return error.response(err, res);
     }
 
     return res.status(utils.http.CREATED).location(`${req.baseUrl}/items/${new_item_id}`).json({ id: new_item_id });
@@ -57,13 +52,13 @@ async function createItem(req, res) {
     Gestisce la ricerca di item secondo criteri vari
 */
 async function searchItem(req, res) {
+    let items = [];
     let query_criteria = {};
-    let sort_criteria = { "relevance": -1 }
+    let sort_criteria = { "relevance": -1 };
 
     // Composizione della query
     if (req.query.category) {
         const category = await CategoryModel.findByName(req.query.category);
-        if (!category) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(MISSING_CATEGORY_MESSAGE)); }
         query_criteria.category_id = category._id; 
     }
     if (req.query.name) { query_criteria.name = `/${req.query.name}/`; }
@@ -74,33 +69,37 @@ async function searchItem(req, res) {
     if (req.query.name_asc) { sort_criteria = { "name": 1 }; }
     if (req.query.name_desc) { sort_criteria = { "name": -1 }; }
 
-    const items = await ItemModel.aggregate([
-        { $match: query_criteria },
-        { $lookup: {
-            from: ProductModel.collection.name,
-            localField: "products_id",
-            foreignField: "_id",
-            as: "products"
-        }},
-        { $project: {
-            name: 1,
-            description: 1,
-            min_price: { $min: "$products.price" }, // Il prezzo che rappresenta l'item è quello più piccolo tra i suoi prodotti
-            product_number: { $size: "$products" },
-            image_path: { // L'immagine associata ad un item è la prima del primo prodotto
-                $first: {
-                    $getField: { field: "images_path", input: { $first: "$products" } }
+    try {
+        items = await ItemModel.aggregate([
+            { $match: query_criteria },
+            { $lookup: {
+                from: ProductModel.collection.name,
+                localField: "products_id",
+                foreignField: "_id",
+                as: "products"
+            }},
+            { $project: {
+                name: 1,
+                description: 1,
+                min_price: { $min: "$products.price" }, // Il prezzo che rappresenta l'item è quello più piccolo tra i suoi prodotti
+                product_number: { $size: "$products" },
+                image_path: { // L'immagine associata ad un item è la prima del primo prodotto
+                    $first: {
+                        $getField: { field: "images_path", input: { $first: "$products" } }
+                    }
                 }
-            }
-        }},
-        { $sort: sort_criteria },
-        { $skip: parseInt(req.query.page_number) }, // Per paginazione
-        { $limit: parseInt(req.query.page_size) }
-    ]).catch (function (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
-    });
+            }},
+            { $sort: sort_criteria },
+            { $skip: parseInt(req.query.page_number) }, // Per paginazione
+            { $limit: parseInt(req.query.page_size) }
+        ]);
 
-    if (items.length === 0) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (items.length === 0) { throw error.generate.NOT_FOUND("Nessun prodotto corrisponde ai criteri di ricerca"); }
+    }
+    catch (err) {
+        return error.response(err, res);
+    }
+
     return res.status(utils.http.OK).json(items);
 }
 
@@ -113,15 +112,14 @@ async function searchItemByBarcode(req, res) {
     try {
         // Cerca il prodotto associato al barcode e poi l'item che contiene il prodotto
         const product = await ProductModel.findOne({ barcode: req.params.barcode }).exec();
-        if (product) {
-            item = await ItemModel.findOne({ products_id: product._id }).exec();
-        }
+        if (product) { item = await ItemModel.findOne({ products_id: product._id }).exec(); }
+
+        if (!item) { throw error.generate.NOT_FOUND("Nessun item associato al barcode"); }
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR); 
+        return error.response(err, res);
     }
     
-    if (!item) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
     return res.status(utils.http.OK).json(item);
 }
 
@@ -133,11 +131,11 @@ async function searchProducts(req, res) {
 
     try {
         const item = await ItemModel.findById(req.params.item_id).populate("products_id", "-__v").exec();
-        if (!item) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!item) { throw error.generate.NOT_FOUND("Item inesistente"); }
         products = item.products_id;
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.status(utils.http.OK).json(products);
@@ -152,19 +150,18 @@ async function updateItemById(req, res) {
     // Estrazione id della categoria
     if (updated_fields.category) {
         const category = await CategoryModel.findByName(updated_fields.category);
-        if (!category) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(MISSING_CATEGORY_MESSAGE)); }
         delete updated_fields.category;
         updated_fields.category_id = category._id;
     }
 
     try {
         const updated_item = await ItemModel.findByIdAndUpdate(req.params.item_id, updated_fields, { new: true });
-        if (!updated_item) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!updated_item) { throw error.generate.NOT_FOUND("Item inesistente"); }
     
         return res.status(utils.http.OK).json(updated_item);
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 }
 
@@ -178,12 +175,13 @@ async function updateProductByIndex(req, res) {
     // Estrazione del prodotto a partire dall'item
     try {
         const item = await ItemModel.findById(req.params.item_id, { "products_id": 1 }).exec();
-        if (!item || !item.products_id[parseInt(req.params.product_index)]) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!item) { throw error.generate.NOT_FOUND("Item inesistente"); }
+        if (!item.products_id[parseInt(req.params.product_index)]) { throw error.generate.NOT_FOUND("Prodotto inesistente"); }
     
         updated_product = await ProductModel.findByIdAndUpdate(item.products_id[parseInt(req.params.product_index)], updated_fields, { new: true });
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.status(utils.http.OK).json(updated_product);
@@ -196,7 +194,7 @@ async function deleteItemById(req, res) {
     try {
         // Estrazione dei prodotti associati all'item
         const to_delete_item = await ItemModel.findById(req.params.item_id, { products_id: 1 }).exec();
-        if (!to_delete_item) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!to_delete_item) { throw error.generate.NOT_FOUND("Item inesistente"); }
 
         // Rimozione delle immagini associate ai prodotti
         const products = await ProductModel.find({ _id: { $in: to_delete_item.products_id } }, { images_path: 1 }).exec();
@@ -211,7 +209,7 @@ async function deleteItemById(req, res) {
         await ItemModel.findByIdAndDelete(req.params.item_id);
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.sendStatus(utils.http.OK);
@@ -224,7 +222,8 @@ async function deleteProductByIndex(req, res) {
     try {
         // Estrazione del prodotto
         const item = await ItemModel.findById(req.params.item_id, { products_id: 1 }).exec();
-        if (!item || !item.products_id[parseInt(req.params.product_index)]) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!item) { throw error.generate.NOT_FOUND("Item inesistente"); }
+        if (!item.products_id[parseInt(req.params.product_index)]) { throw error.generate.NOT_FOUND("Prodotto inesistente"); }
         if (item.products_id.length === 1) { 
             return res.status(utils.http.SEE_OTHER).location(`${req.baseUrl}/items/${req.params.item_id}`).json({message: "Per rimuovere questo prodotto bisogna rimuovere l'item"}) 
         }
@@ -240,7 +239,7 @@ async function deleteProductByIndex(req, res) {
         await ItemModel.findByIdAndUpdate(req.params.item_id, { $pull: { products_id: to_delete_product._id } });
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.sendStatus(utils.http.OK);
@@ -254,9 +253,10 @@ async function createUploadImages(req, res) {
     try {
         // Ricerca dell'id del prodotto
         const item = await ItemModel.findById(req.params.item_id, { products_id: 1 }).exec();
-        if (!item) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!item) { throw error.generate.NOT_FOUND("Item inesistente"); }
+        if (!item.products_id[parseInt(req.params.product_index)]) { throw error.generate.NOT_FOUND("Prodotto inesistente"); }
+
         const product_id = item.products_id[parseInt(req.params.product_index)];
-        if (!product_id) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
 
         // Salvataggio dei file nel filesystem
         let files_name = []
@@ -271,7 +271,7 @@ async function createUploadImages(req, res) {
         await ProductModel.findByIdAndUpdate(product_id, { $push: { images_path: { $each: files_name } } });
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.sendStatus(utils.http.OK);
@@ -282,11 +282,15 @@ async function createUploadImages(req, res) {
 */
 async function deleteImageByIndex(req, res) {
     try {
-        // Estrazione del prodotto
+        // Estrazione dell'item
         const item = await ItemModel.findById(req.params.item_id, { products_id: 1 }).exec();
-        if (!item || !item.products_id[parseInt(req.params.product_index)]) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!item) { throw error.generate.NOT_FOUND("Item inesistente"); }
+        if (!item.products_id[parseInt(req.params.product_index)]) { throw error.generate.NOT_FOUND("Prodotto inesistente"); }
+        
+        // Estrazione del prodotto
         const product = await ProductModel.findById(item.products_id[parseInt(req.params.product_index)], { images_path: 1 }).exec();
-        if (!product || !product.images_path[parseInt(req.params.image_index)]) { return res.status(utils.http.NOT_FOUND).json(error.formatMessage(utils.http.NOT_FOUND)); }
+        if (!product) { throw error.generate.NOT_FOUND("Prodotto inesistente"); }
+        if (!product.images_path[parseInt(req.params.image_index)]) { throw error.generate.NOT_FOUND("Immagine inesistente"); }
 
         // Determina l'immagine da cancellare
         const to_delete_image = product.images_path[parseInt(req.params.image_index)];
@@ -296,7 +300,7 @@ async function deleteImageByIndex(req, res) {
         await ProductModel.findByIdAndUpdate(product._id, { $pull: { images_path: to_delete_image } });
     }
     catch (err) {
-        return res.sendStatus(utils.http.INTERNAL_SERVER_ERROR);
+        return error.response(err, res);
     }
 
     return res.sendStatus(utils.http.OK);
